calculate drop rate only once
Problem: The random numbers calculated in the update_position and
update_color shader did not match. The consequence was, that the
repositioned particles had the wrong color.

Solution: The drop is only calculated once in the update_color shader
and the updated colors are passed to the update_position shader. If the
alpha value is zero, the particle is droped and gets randomly
repositioned.